Fixed a few bugs an added some flashing lights (no really ;).
The 2nd last commit with UI changes for help/introduction changes moved about some of the list-adapter set-up which didn't work to well adjusting to changes in the list. These are now fixed.

There is still issues in th ArtefactSettingsActivity when thumbnails are added. The UI does not adjust tofit them it properly.